Main parts of project
The .xlsx is the raw data downloaded and the your_questions.js is automatically generated by excel_to_pdf_fixed.py, which uses PracticeQuestions.xlsx as an experiment. Then use json2pdf.py to convert the JSON file into .tex code, hopefully through which we could compile a PDF. (but it didn't, that is why I am here). As you could see the final output is lack of images (which I will mention in another tree and all the inserted object name is repeated)